builder - skip modules with existing archive, so we can more easily retry failed modules

+++ scriptmodules/admin/builder.sh
@@ -21,15 +21,27 @@ function module_builder() {
     local ids=($@)
 
     local id
-    local mode
     for id in "${ids[@]}"; do
-        [[ "$id" =~ ^[0-9]+$ ]] && id="$(rp_getIdFromIdx $id)"
-        ! fnExists "install_$id" && continue
+        # if index get mod_id from array else we look it up
+        local md_id
+        local md_idx
+        if [[ "$id" =~ ^[0-9]+$ ]]; then
+            md_id="$(rp_getIdFromIdx $id)"
+            md_idx="$id"
+        else
+            md_idx="$(rp_getIdxFromId $id)"
+            md_id="$id"
+        fi
+        ! fnExists "install_${md_id}" && continue
+
+        # skip already built archives, so we can retry failed modules
+        [[ -f "$__tmpdir/archives/$__os_codename/$__platform/${__mod_type[md_idx]}/$md_id.tar.gz" ]] && continue
 
         # build, install and create binary archive.
         # initial clean in case anything was in the build folder when calling
+        local mode
         for mode in clean remove depends sources build install create_bin clean remove "depends remove"; do
-            rp_callModule "$id" $mode
+            rp_callModule "$md_id" $mode
             # return on error
             [[ $? -eq 1 ]] && return 1
             # no module found - skip to next module
